Refactor: split filter builders into submodules
